requirements.txt
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
fonttools==4.61.1
graphviz==0.21
influxdb-client==1.49.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
pillow==12.0.0
pycparser==2.23
pydyf==0.12.1
pyparsing==3.3.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
reactivex==4.1.0
reportlab==4.4.7
setuptools==80.9.0
six==1.17.0
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
